Regional_Cybersec_Chatbot_UK_Malta-main/UK/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
beautifulsoup4==4.12.3
chromadb==0.4.22
sentence-transformers==2.2.2
langchain==0.1.0
langchain-ollama==0.0.2
pypdf2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
aiofiles==23.2.1
